# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway of Restaurant Reservation System",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/reservation-system/menu": {
            "get": {
                "description": "Retrieves multiple meals info from menu table in PostgreSQL",
                "tags": [
                    "menu"
                ],
                "summary": "Fetches meals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Number of meals to fetch",
                        "name": "limit",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Number of meals to omit",
                        "name": "offset",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.Meals"
                        }
                    },
                    "400": {
                        "description": "Invalid pagination parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while fetching meals from menu",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts new meal info to menu table in PostgreSQL",
                "tags": [
                    "menu"
                ],
                "summary": "Adds a meal to menu",
                "parameters": [
                    {
                        "description": "New data",
                        "name": "new_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menu.MealDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.ID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while adding meal to menu",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation-system/menu/{meal_id}": {
            "get": {
                "description": "Retrieves meal info from menu table in PostgreSQL",
                "tags": [
                    "menu"
                ],
                "summary": "Gets a meal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meal ID",
                        "name": "meal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.MealInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid meal ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while getting meal from menu",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates meal info in menu table in PostgreSQL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "Updates a meal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meal ID",
                        "name": "meal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New info",
                        "name": "new_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menu.MealInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid meal ID or data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while updating meal in menu",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes meal info from menu table in PostgreSQL",
                "tags": [
                    "menu"
                ],
                "summary": "Deletes a meal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meal ID",
                        "name": "meal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid meal ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while removing meal from menu",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation-system/payments": {
            "post": {
                "description": "Inserts new payment info to payments table in PostgreSQL",
                "tags": [
                    "payment"
                ],
                "summary": "Creates a payment",
                "parameters": [
                    {
                        "description": "New data",
                        "name": "new_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.PaymentDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment.Status"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while creating payment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reservation-system/payments/{payment_id}": {
            "get": {
                "description": "Retrieves payment info from payments table in PostgreSQL",
                "tags": [
                    "payment"
                ],
                "summary": "Gets a payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment.PaymentInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid payment ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while getting payment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates payment info in payments table in PostgreSQL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Updates a payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New info",
                        "name": "new_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment.PaymentInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid payment ID or data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while updating payment",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "menu.ID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "menu.MealDetails": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
        "menu.MealInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
        "menu.Meals": {
            "type": "object",
            "properties": {
                "meals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/menu.MealInfo"
                    }
                }
            }
        },
        "payment.PaymentDetails": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "payment_method": {
                    "type": "string"
                },
                "reservation_id": {
                    "type": "string"
                }
            }
        },
        "payment.PaymentInfo": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "reservation_id": {
                    "type": "string"
                }
            }
        },
        "payment.Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}